% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataGenerator}
\alias{dataGenerator}
\title{dataGenerator}
\usage{
dataGenerator(n, beta = c(0, 1, 2), sigma = 1)
}
\arguments{
\item{n}{number of data points}

\item{beta}{parameter vector}

\item{sigma}{standard deviation of the Gaussian noise}
}
\value{
a list containing the response vector Y (size n), a dataframe for explanatory variables X (size nxp) and the vector of Gaussian noise (size n)
}
\description{
Generate data for multiple linear regression (using the standard uniform distribution)
}
\examples{
dataset <- dataGenerator(10, 1:3, 2)
}
