% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcls2.R
\name{pcls2}
\alias{pcls2}
\title{pcls2}
\usage{
pcls2(X, Y)
}
\arguments{
\item{X}{a dataframe containing the explanatory variables (2 columns)}

\item{Y}{a dataframe containing the response vector}
}
\value{
a list containing the optimal parameter vector and the vector of Lagrange multipliers
}
\description{
Solve the positively constrained least squares problem with 2 explanatory variables.
}
\examples{
results <- pcls2(X = data.frame(runif(3), runif(3)), Y = data.frame(1:3))
}
