% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcls2_bias_variance_MSE.R
\name{pcls2_bias_variance_MSE}
\alias{pcls2_bias_variance_MSE}
\title{pcls2_bias_variance_MSE}
\usage{
pcls2_bias_variance_MSE(X, beta = c(0, 1), sigma = 1)
}
\arguments{
\item{X}{a dataframe containing the explanatory variables (2 columns)}

\item{beta}{parameter vector}

\item{sigma}{standard deviation of the Gaussian noise}
}
\value{
a dataframe containing the bias, the variance and the MSE of each component of the estimator of beta
}
\description{
Calculate the bias, the variance and the MSE of each component of the estimator of beta in the positively constrained least squares problem with 2 explanatory variables.
}
\examples{
results <- pcls2_bias_variance_MSE(X = data.frame(runif(3), runif(3)), beta = c(1,0.5), sigma = 2)
}
